Move makeTitle to UITheme where it belongs
